♻️ [Refactoring]: [BE] create_task を CreateTaskIntent + TaskIndex::plan_create に集約
PR レビュー議論を受け、application 層の usecase 関数 (create_task_usecase) を
廃止し、ドメインの aggregate メソッド TaskIndex::plan_create に planning ロジック
を集約する。IPC 境界の DTO (CreateTaskArgs) とドメイン VO を繋ぐ翻訳層として
CreateTaskIntent を新設し、From<CreateTaskArgs> for CreateTaskIntent で
正規化（Priority::from_ascii_ci / TaskTitle::from_lenient 等）を 1 箇所に集約。

## 変更点

- task/task_index.rs:
  - CreateTaskIntent / CreateTaskOutcome を追加 (domain)
  - TaskIndex::plan_create を追加（旧 create_task_usecase の責務）
  - resolve_target_dir / existing_filenames_in_dir / join_rel_path /
    build_provisional_task を private fn として取り込み（旧 filename.rs から移管）
- task/task_content.rs:
  - TaskContent::from_intent factory を追加（旧 content.rs::build_task_content から移管）
- task/create/args.rs:
  - From<CreateTaskArgs> for CreateTaskIntent を実装（DTO ↔ domain 境界）
- task/create/command.rs:
  - CreateTaskIntent::from(args) → TaskIndex::from(snapshot).plan_create(...) の
    2 ステップに薄める
- task/create.rs:
  - usecase / filename / content モジュール宣言を削除し、application 層を
    args / command / error の 3 つに簡素化

## 削除ファイル

- task/create/usecase.rs / usecase_tests.rs (10 tests)
- task/create/filename.rs / filename_tests.rs (9 tests;
  TaskFileName::from_title は task_file_name_tests.rs で既にカバー)
- task/create/content.rs / content_tests.rs (4 tests;
  from_intent として task_content_tests.rs に移管)

## 新規テスト

- task/task_index_plan_create_tests.rs (11 tests): plan_create の正常系/異常系/境界値
- task/task_content_tests.rs に from_intent 用 6 tests を追加

## 動作変更なし契約

- IPC wire format (CreateTaskArgs / CreateTaskCommandError Display) は維持
- ErrorKind::AlreadyExists 観測値も維持
- 既存 E2E テスト (command_tests.rs 16 件) が全件通る
- 合計 423 tests pass（旧 429 から 10+9+4 削除 + 11+6 追加 = -6 = 423）
